+++ packages/core/src/command/consolidated/search/fhir-resource/__tests__/ingest-lexical.test.ts
@@ -1,11 +1,12 @@
 import { faker } from "@faker-js/faker";
+import { makePatient } from "../../../../../external/fhir/__tests__/patient";
 import { mergeIntoTargetResource } from "../../../../../fhir-deduplication/shared";
 import { makeCondition } from "../../../../../fhir-to-cda/cda-templates/components/__tests__/make-condition";
 import { filterOutPatientAndRemoveDerivedFromExtensions } from "../ingest-lexical";
 
 describe("singest-lexical", () => {
   describe("filterOutPatientAndRemoveDerivedFromExtensions", () => {
-    it(`returns original array when nothing to hydrate`, async () => {
+    it(`removes derived-from extensions from non-Patient resources`, async () => {
       const conditionId = faker.string.uuid();
       const conditionId2 = faker.string.uuid();
       const condition = makeCondition({ id: conditionId });
@@ -24,5 +25,16 @@ describe("singest-lexical", () => {
       expect(extensions).toBeTruthy();
       expect(extensions.length).toEqual(0);
     });
+  });
+
+  it(`filters out Patient resources`, async () => {
+    const patientResource = makePatient();
+    const result = filterOutPatientAndRemoveDerivedFromExtensions(patientResource);
+    expect(result).toBeUndefined();
+  });
+
+  it(`returns undefined for undefined input`, async () => {
+    const result = filterOutPatientAndRemoveDerivedFromExtensions(undefined);
+    expect(result).toBeUndefined();
   });
 });
