+++ packages/core/src/fhir-deduplication/shared.ts
@@ -1,15 +1,9 @@
-import {
-  CodeableConcept,
-  Coding,
-  Extension,
-  Identifier,
-  Period,
-  Resource,
-} from "@medplum/fhirtypes";
+import { CodeableConcept, Coding, Identifier, Period, Resource } from "@medplum/fhirtypes";
 import { errorToString } from "@metriport/shared";
 import dayjs from "dayjs";
 import utc from "dayjs/plugin/utc";
 import _, { cloneDeep } from "lodash";
+import { createExtensionRelatedArtifact } from "../external/fhir/shared/extensions/derived-from";
 import { capture, out } from "../util";
 import { uuidv7 } from "../util/uuid-v7";
 
@@ -20,11 +14,6 @@ const MISSING_ATTR = "-";
 
 const dateFormats = ["datetime", "date"] as const;
 const unknownValues = ["unknown", "unk", "no known"];
-
-const derivedFromType = "derived-from";
-
-export const artifactRelatedArtifactUrl =
-  "http://hl7.org/fhir/StructureDefinition/artifact-relatedArtifact";
 
 export const UNK_CODE = "UNK";
 export const UNKNOWN_DISPLAY = "unknown";
@@ -50,17 +39,6 @@ export function getDateFromString(dateString: string, dateFormat?: "date" | "dat
   } else {
     return date.format("YYYY-MM-DD");
   }
-}
-
-function createExtensionRelatedArtifact(resourceType: string, id: string | undefined) {
-  return {
-    url: artifactRelatedArtifactUrl,
-    valueRelatedArtifact: { type: derivedFromType, display: `${resourceType}/${id}` },
-  };
-}
-
-export function isDerivedFromExtension(extension: Extension | undefined): boolean {
-  return extension?.valueRelatedArtifact?.type === derivedFromType;
 }
 
 //eslint-disable-next-line @typescript-eslint/no-explicit-any
